llama3/replicate_API/requirements.txt
python-dotenv==1.0.1
replicate==0.28.0
